requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.20.0
streamlit>=1.24.0
snowflake-connector-python[pandas]>=3.1.0
